"rats nest" not "rat nests" + helper funcs
