+++ lib/pcb/create-svg-objects-from-pcb-rats-nest/get-source-port-for-pcb.ts
@@ -1,0 +1,9 @@
+import type { AnyCircuitElement } from "circuit-json"
+import { su } from "@tscircuit/soup-util"
+
+export const getSourcePortForPcbPort = (
+  pcbPortId: string,
+  soup: AnyCircuitElement[],
+) => {
+  return su(soup).source_port.getWhere({ pcb_port_id: pcbPortId })
+}
